+++ RiskOfChaos/EffectDefinitions/Character/Player/Items/MonsterItemSteal.cs
@@ -18,8 +18,23 @@ using UnityEngine;
 namespace RiskOfChaos.EffectDefinitions.Character.Player.Items
 {
     [ChaosEffect("monster_item_steal", DefaultSelectionWeight = 0.6f)]
+    [EffectConfigBackwardsCompatibility("Effect: Steal All Player Items")]
     public sealed class MonsterItemSteal : BaseEffect, ICoroutineEffect
     {
+        const int MAX_ITEM_STEAL_COUNT_DEFAULT_VALUE = 2;
+
+        [EffectConfig]
+        static readonly ConfigHolder<string> _maxItemStealCountConfig =
+            ConfigFactory<string>.CreateConfig("Max Item Steal Count", MAX_ITEM_STEAL_COUNT_DEFAULT_VALUE.ToString())
+                                 .Description("The maximum amount of item stacks each enemy can steal from each player")
+                                 .OptionConfig(new InputFieldConfig())
+                                 .Build();
+
+        static readonly ParsedInt32 _maxItemStealCount = new ParsedInt32
+        {
+            ConfigHolder = _maxItemStealCountConfig
+        };
+
         [EffectConfig]
         static readonly ConfigHolder<bool> _ignoreAILendFilter =
             ConfigFactory<bool>.CreateConfig("Ignore AI Item Blacklist", false)
@@ -79,54 +94,68 @@ namespace RiskOfChaos.EffectDefinitions.Character.Player.Items
             return stealer && stealer.hasBody && victim && victim.hasBody && stealer != victim && stealer.teamIndex != victim.teamIndex && !stealer.IsDeadAndOutOfLivesServer();
         }
 
-        List<CharacterItemStealInitializer> _itemStealInitializers;
+        readonly List<SteppedStealController> _activeStealControllers = new List<SteppedStealController>();
 
         public override void OnStart()
         {
+            _activeStealControllers.Capacity = CharacterMaster.readOnlyInstancesList.Count;
+
+            CharacterMaster.onStartGlobal += initializeMasterForStealing;
+        }
+
+        void initializeMasterForStealing(CharacterMaster master)
+        {
+            if (!master)
+                return;
+
+            CharacterBody body = master.GetBody();
+            if (!body || !body.healthComponent || !body.healthComponent.alive)
+                return;
+
+            CharacterItemStealInitializer stealInitializer = new CharacterItemStealInitializer(master);
+
+            PlayerUtils.GetAllPlayerMasters(true).TryDo(playerMaster =>
+            {
+                if (!getStealableItemStacks(playerMaster.inventory).Any() || !canSteal(stealInitializer.Master, playerMaster))
+                    return;
+
+                stealInitializer.StartStealingFrom(playerMaster.inventory);
+            }, Util.GetBestMasterName);
+
+            if (stealInitializer.ItemStealController)
+            {
+                _activeStealControllers.Add(new SteppedStealController(stealInitializer));
+            }
         }
 
         public IEnumerator OnStartCoroutine()
         {
-            _itemStealInitializers = new List<CharacterItemStealInitializer>(CharacterMaster.readOnlyInstancesList.Count);
-            foreach (CharacterMaster master in CharacterMaster.readOnlyInstancesList)
-            {
-                if (master && !master.IsDeadAndOutOfLivesServer() && master.hasBody)
-                {
-                    _itemStealInitializers.Add(new CharacterItemStealInitializer(master));
-                }
-            }
-
-            PlayerUtils.GetAllPlayerMasters(true).TryDo(playerMaster =>
-            {
-                if (playerMaster.hasBody)
-                {
-                    tryStealItemsFrom(playerMaster);
-                }
-            }, Util.GetBestMasterName);
-
-            List<ItemStealController> activeStealControllers = new List<ItemStealController>(_itemStealInitializers.Select(i => i.ItemStealController));
+            CharacterMaster.readOnlyInstancesList.TryDo(initializeMasterForStealing, Util.GetBestMasterName);
 
             int stealIterations = 0;
             float currentStealInterval = 0.3f;
 
-            while (activeStealControllers.Count > 0)
+            while (_activeStealControllers.Count > 0)
             {
-                Util.ShuffleList(activeStealControllers);
+                Util.ShuffleList(_activeStealControllers);
 
-                for (int i = activeStealControllers.Count - 1; i >= 0; i--)
+                for (int i = _activeStealControllers.Count - 1; i >= 0; i--)
                 {
-                    ItemStealController itemStealController = activeStealControllers[i];
-                    if (!itemStealController)
+                    SteppedStealController stealController = _activeStealControllers[i];
+                    if (!stealController.IsValid())
                     {
-                        activeStealControllers.RemoveAt(i);
+                        _activeStealControllers.RemoveAt(i);
                         continue;
                     }
 
-                    itemStealController.StepSteal();
+                    stealController.ItemStealController.StepSteal();
+                    stealController.NumSteps++;
 
-                    if (!itemStealController.inItemSteal)
+                    if (stealController.NumSteps >= _maxItemStealCount.GetValue(MAX_ITEM_STEAL_COUNT_DEFAULT_VALUE) || !stealController.ItemStealController.inItemSteal)
                     {
-                        activeStealControllers.RemoveAt(i);
+                        _activeStealControllers.RemoveAt(i);
+
+                        stealController.OnLastItemStolen();
                     }
 
                     yield return new WaitForSeconds(currentStealInterval);
@@ -138,24 +167,61 @@ namespace RiskOfChaos.EffectDefinitions.Character.Player.Items
                     }
                 }
             }
-        }
 
-        void tryStealItemsFrom(CharacterMaster master)
-        {
-            if (!getStealableItemStacks(master.inventory).Any())
-                return;
-
-            foreach (CharacterItemStealInitializer itemStealInitializer in _itemStealInitializers)
-            {
-                if (canSteal(itemStealInitializer.Master, master))
-                {
-                    itemStealInitializer.StartStealingFrom(master.inventory);
-                }
-            }
+            onEnd();
         }
 
         public void OnForceStopped()
         {
+            onEnd();
+        }
+
+        void onEnd()
+        {
+            CharacterMaster.onStartGlobal -= initializeMasterForStealing;
+        }
+
+        record SteppedStealController(ItemStealController ItemStealController, CharacterMaster Master, int NumSteps)
+        {
+            public int NumSteps = NumSteps;
+
+            public SteppedStealController(CharacterItemStealInitializer initializer) : this(initializer.ItemStealController, initializer.Master, 0)
+            {
+            }
+
+            public bool IsValid()
+            {
+                if (!ItemStealController || !Master)
+                    return false;
+
+                CharacterBody body = Master.GetBody();
+                if (!body || !body.healthComponent || !body.healthComponent.alive)
+                    return false;
+
+                return true;
+            }
+
+            public void OnLastItemStolen()
+            {
+                if (ItemStealController.inItemSteal)
+                {
+                    IEnumerator waitUntilAllOrbsArrivedThenSetNotStealing()
+                    {
+#pragma warning disable Publicizer001 // Accessing a member that was not originally public
+                        yield return new WaitWhile(() => ItemStealController && ItemStealController.stolenInventoryInfos.Any(i => i.hasOrbsInFlight));
+#pragma warning restore Publicizer001 // Accessing a member that was not originally public
+
+                        if (!ItemStealController)
+                            yield break;
+
+#pragma warning disable Publicizer001 // Accessing a member that was not originally public
+                        ItemStealController.inItemSteal = false;
+#pragma warning restore Publicizer001 // Accessing a member that was not originally public
+                    }
+
+                    ItemStealController.StartCoroutine(waitUntilAllOrbsArrivedThenSetNotStealing());
+                }
+            }
         }
 
         class CharacterItemStealInitializer