+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 **?.?.? Changes:**
 
+* All Items Are A Random Item:
+  * Default selection weight: 1 -> 0.8
+
 * Chat Voting:
   * "Random Effect" option is now guaranteed to not be any of the other options in the vote