+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 **?.?.? Changes:**
 
 * Reworked 1 effect:
-  * Drop All Items -> Loose Pockets: Drops a random item from everyone's inventory every 0.8 seconds
+  * Drop All Items -> Loose Pockets: Drops a random item from everyone's inventory every 0.9 seconds
 
 * Decrease Teleporter Charge Rate:
   * Decreased default charge rate reduction: -50% -> -25%